pyproject.toml
[project]
name = "leaf-backend"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "django",
    "python-decouple",
    "djangorestframework",
    "djangorestframework-simplejwt>=5.5.1",
]

[dependency-groups]
dev = [
    "pytest",
    "pytest-django",
    "pytest-cov",
    "ruff",
    "pre-commit",
    "detect-secrets",
    "pyright>=1.1.408",
    "django-stubs>=6.0.1",
    "djangorestframework-stubs>=3.16.8",
]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings"
pythonpath = ["."]
python_files = ["tests.py", "test_*.py", "*_tests.py"]
addopts = "--nomigrations --reuse-db"
markers = [
    "integration: mark a test as an integration test",
    "slow: mark tests that take a long time to run"
]

[tool.ruff]
line-length = 88
target-version = "py312"
exclude = [
    "**/migrations/",
]

[tool.ruff.lint]
select = ["E", "F", "I"]  # pycodestyle, pyflakes, isort

[tool.pyright]
include = ["apps", "config"]
exclude = [
    "**/migrations/**", 
    "**/node_modules", 
    "**/__pycache__", 
    "**/.*", 
    ".venv", 
    "**/test_*.py"
]
typeCheckingMode = "standard"
reportMissingParameterType = "error"
reportUnknownParameterType = "error"

[tool.django-stubs]
django_settings_module = "config.settings"
